Address critical PR review feedback
Security fixes:
- Move collection creation to server action (actions.ts) to protect auth token
- Schema no longer exposed in client-side bundle

Race condition fix:
- Combine three separate theme useEffects into one to ensure predictable
  order of DOM updates (primary color, title, favicon)
